stable 9.4 : push dev changes to stable build
f366c346 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/824, fix newline in changelog
aee30faf BuildBot: translation: update po file automatically
5a4c46af web: update changelog
a289c4b4 web: improve ppp access concentrator data
89255e7b Update webui-core.po (it) (#823)
8caf086c web: adds option to use default openwrt backups and restore functions
6108118a Update webui-core.po (it) (#820)
cf963384 webui: fix js error on reboot and reset functions
beae4b61 webui: avoid concurrent request on command log modal, clear background update ajax threads during reset/upgrades
993dc5c5 string improvement, update it lang
0b166aae improved version check js, fix pppoe internet modal
feaf9574 set blinking led (simulate sysupgrade) while executing rootdevice also fix led remain in blinking status on first install
355db8a0 BuildBot: translation: update po file automatically
009744fa remove useless defs expose concentrator and ipv6
53a6dedc fix few data update in wan stats, expose concentrator name/MAC when connected in PPP
688b1b0b mapper: export concentrator name in ppp session
a70ac032 web: generalize ipv6 addrs and prefix code
0b17946f mapper: fix not detected interface if iface6rd option not set
d8cf32a2 web: fix logout issue
5873f0a7 mapper: rework modgui rpc commitapply
0203377f script: fix typo
ff5ce560 mapper: check cache at every get request
2f318b1b BuildBot: translation: update po file automatically
d200c11f web: change Advanced Reset function
e9550d85 web: fix some errors with ipv6
de7eb71b scripts: fix logic error in app condition
7c07aa93 fix specific ipk installation on first install
f60df797 webui: fix xdsl driver reset
6b972903 webui: improved checkver and xdsl driver change
a535f5a6 Update webui-core.po (it) (#806)
fe8ef087 webui: hide PPP stats when not connected (hide exact path not found) minor string fixup, update changelog
8d9cee83 webui: skip stop/starting Ko nulled intervals (fix Update GUI button after check upgrade clicked)
da4d9431 webui: fix error 500 on xdsl modal save minor fixes to refresh driver script
d9df194c rootdevice: fix logic error in version set and error in circleci scripts
0e65c2b4 circleci: fix small bug
968c8fff BuildBot: translation: update po file automatically
e6141e0e rootdevice: append short git hash to gui version
0ad9f895 web: improve xdsl modal driver install
4f820ba9 web: fix broken autoscroll on command-log modal
94a7b656 web: make ajax update generic for Internet card
bdf75b41 web: improve view of command-log-read
743f3f24 web: fix js error if mobile disabled
42df66b9 webui: fix js checkver
153631ee webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/798 mappers: avoid losing script execution status on transformer restart
4962d886 webui: fix error 500 extension card/modal
f0232711 BuildBot: translation: update po file automatically
4b0e6b57 webui: pppoe IPv6 (#802)
c1a8dd97 latest fixes for https://github.com/Ansuel/tch-nginx-gui/issues/529
f5637f75 moved all password mappers as normal options https://github.com/Ansuel/tch-nginx-gui/issues/529
443ecd0d Update webui-core.po (de) (#803)
c1dddac9 fix https://github.com/Ansuel/tch-nginx-gui/issues/795
2dfe20eb minor upgradegui code reformatting
c7409cfc add new TIM ACS url
90e21861 more fixes for https://github.com/Ansuel/tch-nginx-gui/issues/529
11244337 upgradegui: check saved gui in /root and upgrade if necessary.
61e8ff54 web: fix checver loop
32eca28d various fixes for https://github.com/Ansuel/tch-nginx-gui/issues/529
ed771fda web: fix modal modgui
446d3853 web: use new commandlogread ajax for checkver
bffaf074 fix https://github.com/Ansuel/tch-nginx-gui/issues/407#issuecomment-526900929
ad4ff5de fix https://github.com/Ansuel/tch-nginx-gui/issues/795
392ca6f5 fix background error 500 during upgrade fix https://github.com/Ansuel/tch-nginx-gui/issues/790 improvements to specific pack setup
9c64e570 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/792 https://github.com/Ansuel/tch-nginx-gui/issues/407#issuecomment-526854150
7fade410 Update webui-core.po (it) (#791)
684d1cbf webui: fix progress points and Idle status restore when script execution terminated
c3e6b71f webui: avoiding connection issue popup during upgrade, disable autorefresh if login expire during command execution
a7bb0ba2 web: skip delete of commandlogread
f2148f9a scripts: fix error in postreq
751f2160 web: rework modgui modal and use new log modal
5350d05d webui: improve command log modal
13187591 webui: adds support for stop and restart KoRequest
2b083735 rootdevice: drop use of transformer for logging
6a7fbabd mapper: use executeCommand instead of upgradegui mapper
1893c768 mapper: set requested in wrapper instead of specific script
a5657c59 mapper: rework upgradegui/scriptRequest logic
f9acf981 webui: fix wrapper script permission
9b2f2a95 BuildBot: translation: update po file automatically
95ff5297 webui: fix update, show script execution status while installing app, minor refactoring
9311a83d BuildBot: translation: update po file automatically
6130ed7e web: initial support for log export in gui
e115ea40 web: fix bug with toggle modal and adds function to load modal manually
8a722826 add xdsl drivers A2pvfbH045l B2pvfbH045l (#787)
cd6936d4 web: introduce support for custom refresh function in ajax refresh
233dfe2b Merge pull request #784 from tjeh/master
7e151263 add INWX.COM
56e49c70 add INWX.COM
b9d27705 webui: fix (for real) https://github.com/Ansuel/tch-nginx-gui/issues/772
34635ba8 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/772 modgui_script: fix ssh rootlogin on wan, odhcp wrongly reenabled on each run
77a66a80 modgui_script: move file removal on end of upgrade device as package in tmp file can be searched by install script
feb2947b modgui_script: Remove all GUI package stripping feature for low-flash devices, added Free RAM check when running install script
1b33b06f upgrade: fixed upgrade script for low-ram devices, improved logging modgui_script: improved logging webui: updated changelog
8aca494d rootdevice: save modgui file across firmware update
ad0f84c8 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/778
84fd3929 webui: fix typo for https://github.com/Ansuel/tch-nginx-gui/issues/765
520c619d webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/765
6187d6ce web: various fix to themes
a3ca0fa8 BuildBot: translation: update po file automatically
971f7621 webui: added reset CWMP button
52997ab7 webui: allow force https to be instantly activated
f37e9bb5 modgui-script: fix https://github.com/Ansuel/tch-nginx-gui/issues/759, muted firewall restart output modgui-script: continue flash-blinking after the led service restarted during upgrade webui: revert https://github.com/Ansuel/tch-nginx-gui/commit/063f66f3dcd139faac194aac500beb813446f6e4 webui: deleting also cwmp.db when resetting configs
8517e39c webui: fix typo in eco modal, should fix cannot set wifi nsc https://github.com/Ansuel/tch-nginx-gui/issues/769
7fae4aa0 lang: fix https://github.com/Ansuel/tch-nginx-gui/issues/771
28243cd6 webui: show both ipv4/6 and uptime in dhcp mode
c936cdf8 BuildBot: translation: update po file automatically
1d1bf315 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/766
03369682 minor string optimization
beee1c85 webui: fix https://github.com/Ansuel/tch-nginx-gui/issues/765
e6f1f580 webui: fix unified-ui theme menu hovering on mobile, remove debug log
8a285743 webui: improvements for unified-ui theme
e0977a44 webui: improvements for mobile unified-ui theme
91337e8e lang: fix typo in webui-voice
063f66f3 webui: reload entire config on call of reloadUsers function
fe3393a9 webui: fix modal footer going oversize
13b694fa webui: avoid newline on long modal title
ad76c72c webui: unified ui theme improvements https://github.com/Ansuel/tch-nginx-gui/issues/724
84c165a7 webui: hide dosprotect and flow cache when config not present (device not compatible)
cef220d0 wol: improve init script

+++ decompressed/gui_file/etc/modgui_scripts/02_specific.sh
@@ -43,10 +43,69 @@ extract_with_check() {
   return $RESTART_SERVICE
 }
 
+apply_right_opkg_repo() {
+  logger_command "Checking opkg feeds config"
+	marketing_version="$(uci get -q version.@version[0].marketing_version)"
+
+	opkg_file="/etc/opkg.conf"
+
+	case $marketing_version in
+	"18.3"*)
+		if [ -n "$(  grep $opkg_file -e "brcm63xx-tch" )" ]; then
+			rm /etc/opkg.conf
+			cp /rom/etc/opkg.conf /etc/
+		fi
+		if [ -z "$(  grep $opkg_file -e "Ansuel/GUI_ipk/kernel-4.1" )" ]; then
+			cat << EOF >> $opkg_file
+arch all 100
+arch arm_cortex-a9 200
+arch arm_cortex-a9_neon 300
+src/gz chaos_calmer_base https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/base
+src/gz chaos_calmer_packages https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/packages
+src/gz chaos_calmer_luci https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/luci
+src/gz chaos_calmer_routing https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/routing
+src/gz chaos_calmer_telephony https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/telephony
+src/gz chaos_calmer_core https://raw.githubusercontent.com/Ansuel/GUI_ipk/kernel-4.1/target/packages
+EOF
+		fi
+		;;
+	"17.3"*)
+		if [ -z "$(  grep $opkg_file -e "roleo/public/agtef/1.1.0/brcm63xx-tch" )" ]; then
+			cat << EOF >> $opkg_file
+src/gz chaos_calmer_base https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/base
+src/gz chaos_calmer_packages https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/packages
+src/gz chaos_calmer_luci https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/luci
+src/gz chaos_calmer_routing https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/routing
+src/gz chaos_calmer_telephony https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/telephony
+src/gz chaos_calmer_management https://repository.ilpuntotecnico.com/files/roleo/public/agtef/1.1.0/brcm63xx-tch/packages/management
+EOF
+		fi
+		;;
+	"16.3"*)
+		if [ -z "$(  grep $opkg_file -e "roleo/public/agtef/brcm63xx-tch" )" ]; then
+			cat << EOF >> $opkg_file
+src/gz chaos_calmer_base https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/base
+src/gz chaos_calmer_packages https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/packages
+src/gz chaos_calmer_luci https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/luci
+src/gz chaos_calmer_routing https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/routing
+src/gz chaos_calmer_telephony https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/telephony
+src/gz chaos_calmer_management https://repository.ilpuntotecnico.com/files/roleo/public/agtef/brcm63xx-tch/packages/management
+EOF
+		fi
+		;;
+	*)
+		logger_command "No opkg file supported"
+		;;
+	esac
+}
+
 ledfw_extract() {
   if [ -f "/tmp/ledfw_support-specific$1.tar.bz2" ]; then
     extract_with_check "/tmp/ledfw_support-specific$1.tar.bz2"
-    [ $? -eq 1 ] && /usr/share/transformer/scripts/restart_leds.sh
+    if [ $? -eq 1 ]; then
+      /usr/share/transformer/scripts/restart_leds.sh
+      ubus send fwupgrade '{"state":"upgrading"}' #avoid losing the flashing-state when service restarted
+    fi
   fi
 }
 
@@ -143,7 +202,7 @@ install_specific() {
   if ping -q -c 1 -W 1 8.8.8.8 >/dev/null 2>&1; then
     logger_command "Applying specific model fixes..."
     uci set modgui.app.specific_app="0"
-    /usr/share/transformer/scripts/appInstallRemoveUtility.sh install specific_app "$1"
+    /usr/share/transformer/scripts/appInstallRemoveUtility.sh install specificapp "$1"
     uci set modgui.app.specific_app="1"
   else
     logger_command "No connection detected, install specific upgrade pack manually!"
@@ -152,13 +211,18 @@ install_specific() {
 }
 
 remove_wizard_5ghz() {
-  logger_command "Removing 5GHz config from wizard..."
-  rm /www/wizard-cards/*wireless_5G*
+  if [ -n "$(find /www/wizard-cards/ -iname '*wireless_5G*')" ]; then
+    logger_command "Removing 5GHz config from wizard..."
+    rm /www/wizard-cards/*wireless_5G*
+  fi
 }
 
 #THIS CHECK DEVICE TYPE AND INSTALL SPECIFIC FILE
 device_type="$(uci get -q env.var.prod_friendly_name)"
 kernel_ver="$(< /proc/version awk '{print $3}')"
+
+
+[ -z "${device_type##*DGA413*}" ] && apply_right_opkg_repo #Check opkg conf based on version
 
 if [ ! "$(uci get -q modgui.app.specific_app)" ]; then
   uci set modgui.app.specific_app="0"
@@ -166,8 +230,10 @@ fi
 
 if [ -z "${device_type##*DGA413*}" ]; then
   install_specific DGA
-elif [ -z "${kernel_ver##3.4*}" ] && [ -z "${device_type##*TG789*}" ]; then
+elif [ -z "${kernel_ver##3.4*}" ] && [ -z "${device_type##*TG789*}" ] && [ -n "${device_type##*Xtream*}" ]; then
   install_specific TG789
+elif [ -z "${device_type##*TG789*}" ] && [ -z "${device_type##*Xtream*}" ]; then
+  install_specific TG800 #use this package as it contain only the right telnet binary for arm
 elif [ -z "${kernel_ver##3.4*}" ] && [ -z "${device_type##*TG799*}" ]; then
   install_specific TG789
 elif [ -z "${kernel_ver##3.4*}" ] && [ -z "${device_type##*TG800*}" ]; then
@@ -213,6 +279,6 @@ else
 fi
 
 if [ -f /tmp/custom-ripdrv-specificDGA.tar.bz2 ]; then
-  logger_command "Removing ripdrv and resuming root process..."
+  logger_command "Removing specific packages from tmp..."
   rm /tmp/*specific*.tar.bz2
 fi
